The lexer.go unit tests now pass
Because of look ahead, I current append a 0 byte at the end of the
string. When reading a file, the expectation is that in general the
buffer will be a tad larger than the content, so we will get the 0 byte
for free without copy.

Pass --no-generation-date to re2go to make the output deterministic.