Make Debug and Metrics usable as nin package
This is an awkward functionality, should probably be converted to
something Go native later. For now keeping mostly as-is so the numbers
can match up (at least partially) with what ninja outputs.

Ran:
  gorename -from '"github.com/maruel/nin".MetricRecord' -to metricRecord
  gorename -from '"github.com/maruel/nin".ManifestParserOptions.Explain' -to explain

and a lot more.

With this, nin is finally usable as a library and ninja.go can be moved
out.