Process subninja files at the end
This permits the system to concentrate on one file at a time; instead of
doing depth first.

This does reorder the edges but it shouldn't impact the build process
much.

The rationale is that in a follow up we'd do the processing
asynchronously. Not doing here to keep things simple.