Make code more Go idiomatic
This one was tricky. plan.addSubTarget() has to return a bool.
plan.refreshDyndepDependents() was incorrectly *not* returning a bool, I
had incorrectly changed it in a previous commit. Sadly these functions
do not have function level comments.

While at it, changed DependencyScan.recomputeOutputsDirty() to return a
bool instead of taking a pointer to a bool as output.