Leverage the 5 previous commits to make parsing and processing concurrent
This should have reduces overall latency by using multiple cores instead
of one.
This should have been true especially when using subninja files.

In practice, this is significantly slower. Committing since it's not the
default processor anymore to be able to iterate on this to figure out if
it can be improved.