Preallocate 60000 items of Edge and Node
The idea is to reduce the amount of heap search when allocating memory.

When set to a value too low (let's say 10k), it hurts performance. The
Go memory allocator is surprisingly fast, or array lookups are
surprisingly slow?

At 60k it does help. I was able to get the best Fuchsia no-op
incremental build time ever at 5.278s.

Since the fine tuning is project dependent, as the value needs to be
just a bit above the number of nodes that will be processed, I prefer to
not check this in for now. Maybe later on we could use a hint instead
that it determined at runtime.